# openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  description: API para la gestión de usuarios que consumiran los servicios de la
    aplicación de tipo Netflix
  title: Microservicio de Usuarios de una aplicación de tipo Netflix
  version: 1.0.0
servers:
- url: /
...
paths:
  /user/logout:
    get:
      description: Cerrar la sesión de la cuenta de un usuario registrado con su email
        y su password en la aplicación
      operationId: logout_usuario
      responses:
        "200":
          description: Se ha cerrado sesión con éxito
        "500":
          description: Error del servidor
      summary: Cerrar sesión en el sistema con la cuenta de un usuario
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuario:
    get:
      description: Retorna el conjunto de usuarios registrados en la aplicación
      operationId: get_all_usuarios
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Usuario'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Usuario'
                type: array
          description: Se ha obtenido exitosamente la lista de usuarios
        "500":
          description: Error del servidor
      summary: Obtener la lista de usuarios registrados en la aplicación
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    post:
      description: Crea una nueva cuenta en la aplicación que estará asociada a un
        nuevo usuario
      operationId: add_usuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
          application/xml:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
            application/xml:
              schema:
                $ref: '#/components/schemas/Usuario'
          description: Usuario creado exitosamente
        "400":
          description: Entrada invalida
        "422":
          description: Excepción en la validación
        "500":
          description: Error del servidor
      summary: Añadir un nuevo usuario a la aplicación
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuario/login:
    post:
      description: Iniciar sesión en la aplicación a través de la cuenta de un usuario
        registrado utilizando su email y su password
      operationId: login_usuario
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/loginUsuario_request'
          application/xml:
            schema:
              $ref: '#/components/schemas/loginUsuario_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/loginUsuario_200_response'
            application/xml:
              schema:
                $ref: '#/components/schemas/loginUsuario_200_response'
          description: Se ha iniciado sesión exitosamente
        "400":
          description: Email y/o password proporcionados inválidos
      summary: Iniciar sesión en el sistema con la cuenta de un usuario
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuario/{idUsuario}:
    delete:
      description: Elimina un contenido multimedia específico de la aplicación en
        función del identificador proporcionado
      operationId: delete_usuario
      parameters:
      - description: ID del usuario a borrar
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Usuario eliminado exitosamente
        "400":
          description: ID proporcionado inválido
        "404":
          description: Contenido no encontrado
        "500":
          description: Error del servidor
      summary: Eliminar un usuario específico por su ID
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    get:
      description: Retorna la información de un usuario en función del identificador
        proporcionado
      operationId: get_usuario_by_id
      parameters:
      - description: ID del usuario a devolver
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
            application/xml:
              schema:
                $ref: '#/components/schemas/Usuario'
          description: Se ha obtenido exitosamente los detalles del usuario
        "400":
          description: ID proporcionado inválido
        "404":
          description: Usuario no encontrado
        "500":
          description: Error del servidor
      summary: Obtener un usuario específico por su ID
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    put:
      description: Actualiza la información de un usuario registrado en función del
        identificador proporcionado
      operationId: update_usuario
      parameters:
      - description: ID del usuario a actualizar
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Usuario'
          application/xml:
            schema:
              $ref: '#/components/schemas/Usuario'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Usuario'
            application/xml:
              schema:
                $ref: '#/components/schemas/Usuario'
          description: Se ha actualizado exitosamente los detalles del usuario
        "400":
          description: ID proporcionado inválido
        "404":
          description: Contenido no encontrado
        "422":
          description: Excepción en la validación
        "500":
          description: Error del servidor
      summary: Actualizar un usuario específico por su ID
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuario/{idUsuario}/Perfiles:
    get:
      description: Retorna el conjunto de perfiles asociados a un usuario específico
        de la aplicación en función del identificador proporcionado
      operationId: get_all_perfiles
      parameters:
      - description: ID del usuario del que se obtendrán sus perfiles
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Perfil'
                type: array
            application/xml:
              schema:
                items:
                  $ref: '#/components/schemas/Perfil'
                type: array
          description: Se ha obtenido exitosamente el listado de perfiles del usuario
        "400":
          description: ID proporcionado inválido
        "500":
          description: Error del servidor
      summary: Obtener la lista de los perfiles asociados a un usuario por su ID
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    post:
      description: Añade un nuevo perfil a un usuario en función del identificador
        proporcionado
      operationId: add_perfil
      parameters:
      - description: ID del usuario al que se le añadirá un nuevo perfil
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
          application/xml:
            schema:
              $ref: '#/components/schemas/Perfil'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Perfil creado exitosamente
        "400":
          description: Entrada invalida
        "422":
          description: Excepción en la validación
        "500":
          description: Error del servidor
      summary: Añadir un nuevo perfil a un usuario por su ID
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
  /usuario/{idUsuario}/{nombrePerfil}:
    delete:
      description: Elimina un perfil específico de un usuario registrado en el sistema
        en función del identificador proporcionado
      operationId: delete_perfil
      parameters:
      - description: ID del usuario del que se borrará el perfil
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: Nombre del perfil que se borrará
        explode: false
        in: path
        name: nombrePerfil
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          description: Perfil eliminado exitosamente
        "400":
          description: ID y/o nombrePerfil proporcionados inválido
        "404":
          description: Usuario o perfil no encontrado
        "500":
          description: Error del servidor
      summary: Eliminar un perfil específico por su nombre
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    get:
      description: Retorna los datos de un perfil específico en función del identificador
        proporcionado del usuario
      operationId: get_perfil
      parameters:
      - description: ID del usuario del que se obtendrá el perfil
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: Nombre del perfil que se obtendrá
        explode: false
        in: path
        name: nombrePerfil
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Se ha obtenido exitosamente el perfil indicado
        "400":
          description: ID y/o nombrePerfil no válidos
        "404":
          description: Perfil no encontrado
        "500":
          description: Error del servidor
      summary: Obtener un perfil específico por su nombre
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
    put:
      description: Actualiza un perfil específico de un usuario registrado en el sistema
        en función del identificador proporcionado del usuario
      operationId: update_perfil
      parameters:
      - description: ID del usuario del que se actualizará el perfil
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: Nombre del perfil que se actualizará
        explode: false
        in: path
        name: nombrePerfil
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Perfil'
          application/xml:
            schema:
              $ref: '#/components/schemas/Perfil'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Perfil'
            application/xml:
              schema:
                $ref: '#/components/schemas/Perfil'
          description: Perfil actualizado exitosamente
        "400":
          description: ID y/o nombrePerfil proporcionados inválidos
        "404":
          description: Perfil no encontrado
        "422":
          description: Excepción en la validación
        "500":
          description: Error del servidor
      summary: Actualizar un perfil específico por su nombre
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
...
  /usuario/{idUsuario}/{nombrePerfil}/{ContenidoFavorito}:
    get:
      description: Retorna un contenido multimedia perteneciente al listado de contenidos
        favoritos de un perfil específico perteneciente a un usuario en función de
        su identificador
      operationId: get_contenido_favorito
      parameters:
      - description: ID del usuario del que se obtendrá el contenido de la lista de
          favoritos de uno de sus perfiles
        explode: false
        in: path
        name: idUsuario
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: nombre del perfil del que se obtendrá el contenido de su lista
          de favoritos
        explode: false
        in: path
        name: nombrePerfil
        required: true
        schema:
          type: string
        style: simple
      - description: ID del contenido marcado como favorito que se obtendrá
        explode: false
        in: path
        name: ContenidoFavorito
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contenido'
            application/xml:
              schema:
                $ref: '#/components/schemas/Contenido'
          description: Se ha obtenido el contenido perteneciente a favoritos exitosamente
        "400":
          description: ID y/o nombrePerfil proporcionados inválidos
        "404":
          description: Perfil no encontrado
        "500":
          description: Error del servidor
      summary: Obtener un contenido multimedia marcado como favorito de un perfil
        específico de un usuario
      tags:
      - usuario
      x-openapi-router-controller: openapi_server.controllers.usuario_controller
components:
  schemas:
    Usuario:
      example:
        metodoPago: paypal
        password: password1
        idUsuario: 5
        perfilesUsuario:
        - favoritosPerfil:
          - 10
          - 10
          nombrePerfil: Marcos1
        - favoritosPerfil:
          - 10
          - 10
          nombrePerfil: Marcos1
        email: asee@gmail.com
        status: activo
      properties:
        idUsuario:
          description: Identificador único del usuario
          example: 5
          title: idUsuario
          type: integer
        email:
          description: Correo electrónico asociado a la cuenta perteneciente al usuario
            y junto a la contraseña permiten el acceso a la misma y a los servicios
            que ofrece la aplicación
          example: asee@gmail.com
          title: email
          type: string
        password:
          description: Contraseña que junto al correo electrónico se encuentran asociados
            a la cuenta perteneciente al usuario y le permiten el acceso a la misma
            y a los servicios que ofrece la aplicación
          example: password1
          title: password
          type: string
        metodoPago:
          description: Forma en la que el usuario realiza el pago correspondiente
            a los servicios contratados en la aplicación
          enum:
          - tarjeta de credito
          - tarjeta virtual
          - tarjeta prepago
          - paypal
          example: paypal
          title: metodoPago
          type: string
        status:
          description: Estado en el que la cuenta asociada al usuario se encuentra
            en este momento
          enum:
          - activo
          - en suspension
          - pendiente de pago
          example: activo
          title: status
          type: string
        perfilesUsuario:
          description: "Conjunto de perfiles pertenecientes a la cuenta asociada al\
            \ usuario. Este atributo se ha definido como un array de objetos, en el\
            \ cual cada posición almacenará la información referente a un perfil,\
            \ es decir, nombrePerfil y favoritosPerfil"
          items:
            $ref: '#/components/schemas/Perfil'
          title: perfilesUsuario
          type: array
      title: Usuario
      type: object
      xml:
        name: usuario
...
    loginUsuario_request:
      properties:
        email:
          description: Correo electrónico asociado a la cuenta perteneciente al usuario
            y junto a la contraseña permiten el acceso a la misma y a los servicios
            que ofrece la aplicación
          title: email
          type: string
        password:
          description: Contraseña que junto al correo electrónico se encuentran asociados
            a la cuenta perteneciente al usuario y le permiten el acceso a la misma
            y a los servicios que ofrece la aplicación
          title: password
          type: string
      title: loginUsuario_request
      type: object
...